pyproject.toml
[project]
name = "rlmc"
version = "1.2.3"
description = "Python utils for AI projects 🚀"
authors = [
    {name = "RayLam2022",email = "1027196450@qq.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
packages = [
    { include = "rlmc",exclude=["*__pycache__*","*ipynb_checkpoints*","*vscode*","","build.*","dist.*","*egg-info","*.mypy_cache","*.git.*"] }
]
include = [
    { path = "medusa-example.ini" }
]
requires-python = ">=3.12,<4"
dependencies = [
    "rich (>=13.9.4,<14.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "psutil (>=6.1.1,<7.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "huggingface-hub (>=0.27.1,<0.28.0)",
    "zhconv (>=1.4.3,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "pynput (>=1.7.7,<2.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "pyperclip (>=1.9.0,<2.0.0)",
    "numpy (>1.26.0,<=2.0.2)",
    "tifffile (>=2025.1.10,<2026.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pillow (<11.0)",
    "h5py (>=3.12.1,<4.0.0)",
    "pypinyin (>=0.53.0,<0.54.0)",
    "zhon (>=2.1.1,<3.0.0)",
    "gitpython (>=3.1.44,<4.0.0)",
]


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# paddle = [
#     "paddlepaddle",
#     "paddlehub"
# ]

# jax=[
#     "jax",
#     "jaxlib"
# ]

# handcontrol=[
#     "pycaw",
#     "mediapipe",  # numpy要小于2.0，与其他冲突，handcontrol要另外处理版本问题
#     "comtypes",
# ]

# game=[
#     "pygame",
#     "gym",
# ]

# tensorflow = [
#     "tensorflow-gpu",
#     "keras"
# ]


[tool.poetry.group.dev.dependencies]
    # "build",
    # "twine"
pytest = "^8.3.4"
mypy = "^1.13.0"


[tool.poetry.group.download.dependencies]
modelscope = "^1.22.2"
codewithgpu = "^0.2.8"


[tool.poetry.group.web.dependencies]
# gradio>4.0.0
beautifulsoup4 = "^4.12.3"
aiofiles = "^24.1.0"
aiohttp = "^3.11.11"
jupyterlab = "^4.3.4"
selenium = "^4.27.1"
fastapi = "^0.115.6"


[tool.poetry.group.video.dependencies]
av = "^14.0.1"
pyaudio = "^0.2.14"
wave = "^0.0.2"
srt = "^3.5.3"
moviepy = "^2.1.2"


[tool.poetry.group.chart.dependencies]
matplotlib = "^3.10.0"
pyecharts = "^2.0.7"
streamlit = "^1.41.1"
streamlit-echarts = "^0.4.0"
fastexcel = "^0.12.1"
polars = "^1.20.0"
hvplot = "^0.11.2"
xlsxwriter = "^3.2.0"


[tool.poetry.group.assistant.dependencies]
# poetry not supporting PEP 517 builds 只能pip install librosa
torch = {path = "D:/torch-2.5.0+cu124-cp312-cp312-win_amd64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu124/torchvision-0.20.0%2Bcu124-cp312-cp312-win_amd64.whl#sha256=e0c79e00469a6ee4982e52d86ea18a29a61e583cce07ba75efd11917dc8d6925"}
torchaudio = {url = "https://download.pytorch.org/whl/cu124/torchaudio-2.5.0%2Bcu124-cp312-cp312-win_amd64.whl#sha256=0ce88068a8880da5cad7bff3cc28a12d8f797e597bacb9a292e1f2f8164ff0cb"}
xformers = {url = "https://download.pytorch.org/whl/cu124/xformers-0.0.28.post2-cp312-cp312-win_amd64.whl#sha256=a8f52e4519640f58c92a2b1e1ad0e1c479b8406eb50f56920d97c3143039f0d2"}
supervision = "<0.24"  # 会安装opencv-python-headless，要卸载opencv-python，再重装opencv-python
pyttsx3 = "^2.98"
qianfan = "^0.4.12.2"
duckduckgo-search = "^7.2.1"
pyaudio = "^0.2.14"
chromadb = "^0.6.3"
pdfplumber = "^0.11.5"
accelerate = "^1.2.1"
transformers = "^4.48.0"
langchain = "^0.3.14"
langchain-community = "^0.3.14"
networkx = "^3.4.2"
timm = "^1.0.13"
einops = "^0.8.0"
segmentation-models-pytorch = "^0.4.0"
torchsummary = "^1.5.1"
thop = "^0.1.1.post2209072238"


[[tool.poetry.scripts]]
mic= { reference = "rltools.mic:main", type = "console" }


